requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
astroid==3.2.4
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
check-swear==0.1.4
click==8.1.7
colorama==0.4.6
dill==0.3.8
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.6
isort==5.13.2
joblib==1.3.2
magic-filter==1.0.12
mccabe==0.7.0
multidict==6.0.5
nltk==3.8.1
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
pydantic==2.5.3
pydantic_core==2.14.6
pylint==3.2.6
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzbar==0.1.9
regex==2024.5.15
requests==2.31.0
scikit-learn==1.4.0
scipy==1.13.0
six==1.16.0
soupsieve==2.5
tabulate==0.9.0
threadpoolctl==3.5.0
tomlkit==0.13.0
tqdm==4.66.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
